requirements.txt
# Core Python environment
pandas==2.1.4
pyarrow==16.1.0

# PySpark for batch ETL
pyspark==3.5.1

# Google Cloud SDK and clients
google-cloud-storage>=2.16.0,<3.0.0
google-cloud-bigquery>=3.17.0,<3.30.0
google-cloud-core>=2.4.1,<3.0.0

# dbt for analytics engineering with BigQuery adapter
dbt-core==1.7.9
dbt-bigquery==1.7.9

# Data quality & observability
great-expectations==1.1.1
elementary-data==0.12.0

# Workflow utilities and logging
pyyaml==6.0.2
rich==13.7.1
tenacity==8.2.3

# Optional: used inside Composer DAGs / local tests
apache-airflow-client==2.9.0

# Testing & formatting (optional but professional)
pytest==8.2.0
black==24.4.2

pydantic==1.10.13

python-dotenv==1.0.1